// packages/shared/src/config/env.d.ts
export declare const config: {
    database: {
        url: string | undefined;
        host: string | undefined;
        port: string | undefined;
        name: string | undefined;
        user: string | undefined;
        password: string | undefined;
    };
    server: {
        port: string;
        nodeEnv: string;
    };
};